requirements.txt
flask
python-dotenv
langchain_huggingface
gunicorn